+++ server/lib/drivers/codec/src/factory_tests.rs
@@ -15,6 +15,7 @@ impl ContentCodec for NullCodec {
             metadata: CodecMetadata::new(ContentType::new("text/utf-8")),
             lossy: false,
             readonly: false,
+            truncated: false,
         })
     }